+++ packages/_core-react/src/App.tsx
@@ -32,7 +32,6 @@ const App = () => {
 		optionsLimit : 5,
 	}
 	const Filters                      = Comp.Filters as React.FC<ComponentFilters>
-
 	return (
 		<>
 			<div>