+++ packages/_core-react/vite.config.ts
@@ -5,15 +5,14 @@
  * @see https://vitejs.dev/config/
  */
 
-import { defineConfig }      from 'vite'
-import react                 from '@vitejs/plugin-react'
-import { svelte }            from '@sveltejs/vite-plugin-svelte'
-import type { PluginOption } from 'vite'
+import { defineConfig } from 'vite'
+import react            from '@vitejs/plugin-react'
+import { svelte }       from '@sveltejs/vite-plugin-svelte'
 
 export default defineConfig( {
 	plugins : [
 		react(),
-		svelte( ) as PluginOption,
+		svelte(),
 	],
 	build : {
 		outDir : 'dist/app',