+++ packages/web/src/lib/components/content/desc/main.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+    
+    import { locales } from "../locales/main";
+
+</script>
+
+<div class="rounded-lg w-full py-8 flex flex-col items-center justify-center max-w-[600px]">
+    <h1 class="font-bold text-[40px] text-primary-500">{$locales.title}</h1>
+    <p>
+        {$locales.description}
+    </p>
+</div>