+++ packages/web/src/lib/components/langSelect.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    
+
     import { t, locales, locale } from '$lib/i18n';
     import { page } from '$app/stores';
     import { goto } from '$app/navigation';
@@ -9,13 +9,13 @@
             goto(target.value)
     }
     
-    $: ({ route } = $page.data);
+    $: ({ route } = $page.data)
 
 </script>
 
 <select 
     on:change={onChange}
-    class="bg-transparent border-0 py-0"
+    class="bg-transparent border-0 py-0 text-[14px] mt-[-2px] pl-0"
     style="box-shadow:none !important"
 >
     {#each $locales as lc}