+++ packages/web/src/lib/style-class.ts
@@ -8,9 +8,11 @@ const pageContent    = 'm-8 flex items-center justify-center'
 const pageContentRow = pageContent + ' flex-row'
 const pageContentCol = pageContent + ' flex-col'
 const fontColor      = 'text-gray-800 dark:text-gray-50'
+const sectionBorder  = 'rounded-xl border-gray-800 border-[1px]'
 
 export const styleClass = {
 	pageContentRow,
 	pageContentCol,
 	fontColor,
+	sectionBorder,
 }