+++ packages/web/src/routes/+layout.ts
@@ -4,24 +4,24 @@
  * @description TODO.
  */
 import { addTranslations, setLocale, setRoute } from '$lib/i18n'
-import { styleClass, type ComponentFilters }    from '$lib'
+import { styleClass }                           from '$lib'
 import type { LayoutRouteId }                   from './$types.js'
 
-const defaultRoute = 	{
-	route : '/',
-	name  : 'Home',
-	title : 'Sidebar Filters',
-}
-const errortRoute  = {
+// const defaultRoute = 	{
+// 	route : '/',
+// 	name  : 'Home',
+// 	title : 'Sidebar Filters',
+// }
+const errortRoute = {
 	route : '/',
 	name  : 'Error page',
 	title : 'Error Page',
 }
 
 const routes = [
-	defaultRoute,
+	// defaultRoute,
 	{
-		route : '/docs',
+		route : MAIN_PKG.extra.docsUrl,
 		name  : 'Documentation',
 		title : 'Documentation',
 	},
@@ -45,38 +45,11 @@ export const load = async ( { data, route } ) => {
 	await setRoute( i18nRoute )
 	await setLocale( lang )
 
-	const filterID                                    = 'pp-super8-custom'
-	const filterOpts: Omit<ComponentFilters, 'type' > = {
-		active : true,
-		id     : 'ps-s8-styles',
-		btn    : { 
-			title : 'Super8',
-		},
-		// content:{
-		//   allowedItems: [
-		//     allowedFilterTabs.custom,
-		//     allowedFilterTabs.preset
-		//   ],
-		// },
-		content     : undefined,
-		locales     : undefined,
-		toSelectors : [ 
-			'#' + filterID,
-		],
-		document : undefined,
-		// optionsLimit : 5,
-	}
-
 	return {
 		...i18n,
 		...getRouteObject( route.id ),
 		routes,
 		classes : styleClass,
-		iframe  : {
-			src : 'https://www.youtube.com/embed/wdTo_Kihcyw?si=ft5lmO6GP3GXi6e_',
-		},
-		filterID,
-		filterOpts,
 	}
 
 }