+++ packages/_core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @s-8/core
 
+## 0.0.3
+
+### Patch Changes
+
+- "@s-8/core": Export flowbite ui library from core
+  "@s-8/docs": Realeas beta version
+  "@s-8/web": Realeas beta version
+
 ## 0.0.2
 
 ### Patch Changes